Add type check on body when POSTing
Add a check on the typeof the body (e.g. when POSTing) to prevent the
client from sending plaintext when `options.json === true`. This is
added to avoid the situation where integrating code attempts to POST
stringified json, which will likely be rejected by the endpoint (because
the payload will appear as plaintext but the `Content-Type` header will
declare 'application/json'.

Also adds tests to confirm plaintext bodies throw errors (but null
payloads are accepted).